disk.h wasn't handling the S2 byte in the FCB properly. The high bit in S2 is used as a "file unmodified" flag. Normally, the BDOS would set this when a file was opened, and the write routines would reset it. RunCPM just pretty much left the flag alone. which was fine, until NULU set it at some point. The next time that file was written to, the code saw an out of range S2 value and a disk full error occured. Imported the changes I'd made to S2 in my ZSDOS'ified disk.h to handle the issue.

+++ RunCPM/disk.h
@@ -284,7 +284,8 @@ uint8 _OpenFile(uint16 fcbaddr) {
 			len = _FileSize(fcbaddr) / BlkSZ;	// Compute the len on the file in blocks
 
 			F->s1 = 0x00;
-			F->s2 = 0x00;
+			F->s2 = 0x80;	// set unmodified flag
+
 
 			F->rc = len > MaxRC ? MaxRC : (uint8)len;
 			for (i = 0; i < 16; ++i)	// Clean up AL
@@ -302,13 +303,15 @@ uint8 _CloseFile(uint16 fcbaddr) {
 	uint8 result = 0xff;
 
 	if (!_SelectDisk(F->dr)) {
-		if (!RW) {
-			_FCBtoHostname(fcbaddr, &filename[0]);
-			if (fcbaddr == BatchFCB)
-				_Truncate((char*)filename, F->rc);	// Truncate $$$.SUB to F->rc CP/M records so SUBMIT.COM can work
-			result = 0x00;
-		} else {
-			_error(errWRITEPROT);
+		if (!(F->s2 & 0x80)) {					// if file is modified
+			if (!RW) {
+				_FCBtoHostname(fcbaddr, &filename[0]);
+				if (fcbaddr == BatchFCB)
+					_Truncate((char*)filename, F->rc);	// Truncate $$$.SUB to F->rc CP/M records so SUBMIT.COM can work
+				result = 0x00;
+			} else {
+				_error(errWRITEPROT);
+			}
 		}
 	}
 	return(result);
@@ -326,7 +329,7 @@ uint8 _MakeFile(uint16 fcbaddr) {
 			if (_sys_makefile(&filename[0])) {
 				F->ex = 0x00;	// Makefile also initializes the FCB (file becomes "open")
 				F->s1 = 0x00;
-				F->s2 = 0x00;
+				F->s2 = 0x00;		// newly created files are already modified
 				F->rc = 0x00;
 				for (i = 0; i < 16; ++i)	// Clean up AL
 					F->al[i] = 0x00;
@@ -451,7 +454,7 @@ uint8 _ReadSeq(uint16 fcbaddr) {
 				F->ex = 0;
 				++F->s2;
 			}
-			if (F->s2 > MaxS2)
+			if ((F->s2 & 0x7F) > MaxS2)
 				result = 0xfe;	// (todo) not sure what to do 
 		}
 	}
@@ -472,6 +475,7 @@ uint8 _WriteSeq(uint16 fcbaddr) {
 			_FCBtoHostname(fcbaddr, &filename[0]);
 			result = _sys_writeseq(&filename[0], fpos);
 			if (!result) {	// Write succeeded, adjust FCB
+				F->s2 &= 0x7F;		// reset unmodified flag
 				++F->cr;
 				if (F->cr > MaxCR) {
 					F->cr = 1;
@@ -506,7 +510,11 @@ uint8 _ReadRand(uint16 fcbaddr) {
 			// adjust FCB unless error #6 (seek past 8MB - max CP/M file & disk size)
 			F->cr = record & 0x7F;
 			F->ex = (record >> 7) & 0x1f;
-			F->s2 = (record >> 12) & 0xff;
+			if (F->s2 & 0x80) {
+				F->s2 = ((record >> 12) & MaxS2) | 0x80;
+			} else {
+				F->s2 = (record >> 12) & MaxS2;
+			}
 		}
 	}
 	return(result);
@@ -527,7 +535,7 @@ uint8 _WriteRand(uint16 fcbaddr) {
 			if (!result) {	// Write succeeded, adjust FCB
 				F->cr = record & 0x7F;
 				F->ex = (record >> 7) & 0x1f;
-				F->s2 = (record >> 12) & 0xff;
+				F->s2 = (record >> 12) & MaxS2;	// resets unmodified flag
 			}
 		} else {
 			_error(errWRITEPROT);
@@ -558,7 +566,7 @@ uint8 _SetRandom(uint16 fcbaddr) {
 
 	int32 count = F->cr & 0x7f;
 	count += (F->ex & 0x1f) << 7;
-	count += F->s2 << 12;
+	count += (F->s2 & MaxS2) << 12;
 
 	F->r0 = count & 0xff;
 	F->r1 = (count >> 8) & 0xff;
